+++ config/routes.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-Assetboar::Engine.routes.draw do
+Assethutch::Engine.routes.draw do
   resources :uploads, only: :create, controller: "direct_uploads" do
     post :complete, on: :member
   end