+++ test/dummy/config/routes.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  mount Assetboar::Engine => "/assetboar"
+  mount Assethutch::Engine => "/assethutch"
 end